+++ flake.nix
@@ -1,9 +1,11 @@
 {
   inputs = {
-    holonix.url = "github:holochain/holochain";
-    holonix.inputs.holochain.url = "github:holochain/holochain/holochain-0.2.0";
-    holonix.inputs.lair.url = "github:holochain/lair/lair_keystore-v0.2.4";
     nixpkgs.follows = "holonix/nixpkgs";
+
+    versions.url = "github:holochain/holochain?dir=versions/0_2";
+    holonix.url = "github:holochain/holochain";
+    holonix.inputs.versions.follows = "versions";
+    holonix.inputs.holochain.url = "github:holochain/holochain/holochain-0.2.1-beta-dev.0";
   };
 
   outputs = inputs@{ holonix, ... }: